refactor:#166 폴더구조변경
